requirements.txt
motor==3.3.2
pydantic==2.5.2
python-dotenv==1.0.0